未完：BookDaoImpl-批量修改、事务处理

+++ src/main/java/com/edut/dao/imp/AccountDaoImpl.java
@@ -19,11 +19,12 @@ implements AccountDao{
 	}
 
 	@Override
-	public void updateBalance(Integer accountId, float amount) {
+	public void updateBalance(Integer accountId, Double amount) {
 		String sql =  " update account_table set "
 					+ " balance = balance - ? "
 					+ " where account_id = ? " ;
 		update(sql, amount , accountId);
 	}
 	
+
 }
